BF: pass /dev/null as stdin upon commands we shouldn't pass stdin into
Closes #1458

Also tests support of -n in sshrun
